requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
celery==4.4.7
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
cryptography==3.2.1
Django==3.1.3
django-celery-results==1.2.1
django-colorfield==0.3.2
django-nose==1.4.7
django-tenant-users==0.3.12
django-tenants==3.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-jwt==1.17.2
kombu==4.6.11
mypy-extensions==0.4.3
nose==1.3.7
parameterized==0.7.4
pathspec==0.8.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
vine==1.3.0
wcwidth==0.2.5
